Initial draft of main view
